+++ custom_addons/einvoice_py/models/__init__.py
@@ -5,6 +5,7 @@ from . import fiscal_py_point_of_issue
 from . import fiscal_py_timbrado
 from . import fiscal_py_csc
 from . import fiscal_py_sequence
+from . import fiscal_py_inutilization
 from . import fiscal_document
 from . import fiscal_document_line
 from . import fiscal_attachment